Transactions

+++ src/database/migrations/20230102183307_accounts.ts
@@ -3,7 +3,7 @@ import { Knex } from 'knex';
 export async function up(knex: Knex): Promise<void> {
   return knex.schema.createTable('accounts', (table) => {
     table.increments('id').primary(),
-      table.bigInteger('balance').defaultTo(0),
+      table.float('balance').defaultTo(0),
       table.string('userEmail').references('email').inTable('users'),
       table.timestamp('createdAt').defaultTo(knex.fn.now());
   });
